+++ lib/main.dart
@@ -9,6 +9,7 @@ import 'package:weight_blade/bloc/weight_entry.dart';
 import 'package:weight_blade/bloc_builders.dart';
 import 'package:weight_blade/event/ledger.dart';
 import 'package:weight_blade/event/reminder.dart';
+import 'package:weight_blade/event/settings.dart';
 import 'package:weight_blade/main_transfer.dart';
 import 'package:weight_blade/repository_builders.dart';
 
@@ -66,6 +67,7 @@ class _MainScreenState extends State<MainScreen> {
     super.initState();
     context.fireEvent<void>(LedgerEvent.loadLedger.event, null);
     context.fireEvent<void>(ReminderEvent.loadReminder.event, null);
+    context.fireEvent<void>(SettingsEvent.loadSettings.event, null);
   }
 
   Widget buildWidget(BuildContext context) {