+++ lib/ui/settings/screen.dart
@@ -1,8 +1,10 @@
 import 'package:event_bloc/event_bloc_widgets.dart';
 import 'package:event_navigation/event_navigation.dart';
 import 'package:flutter/material.dart';
+import 'package:url_launcher/url_launcher.dart';
 import 'package:vhcblade_theme/vhcblade_picker.dart';
 import 'package:vhcblade_theme/vhcblade_widget.dart';
+import 'package:weight_blade/ui/settings/delete.dart';
 
 class SettingsScreen extends StatelessWidget {
   const SettingsScreen({super.key});
@@ -43,13 +45,22 @@ class SettingsPage extends StatelessWidget {
                 style: Theme.of(context).textTheme.headlineMedium),
             const SizedBox(height: 10),
           ]),
+          const ListTile(title: Text("Change Settings"), selected: true),
+          const DeleteConfirmationSettings(),
           ListTile(
               title: const Text("Change Theme"),
               onTap: () => context.fireEvent(
                   NavigationEvent.pushDeepNavigation.event, "theme")),
+          const ListTile(title: Text("Others"), selected: true),
           ListTile(
               title: const Text("Show Licenses"),
               onTap: () => showLicensePage(context: context)),
+          ListTile(
+              title: const Text("Our Other Apps"),
+              onTap: () => launchUrl(
+                    Uri.parse("https://vhcblade.com/#/apps"),
+                    mode: LaunchMode.externalApplication,
+                  )),
         ],
       ),
     );